Fix deletion and try to manage remote file contents. (#81)
Several quality improvements to the Claude logic. Removing menu options
that didn't belong. Tweaking the local cached versions (SyncState) when
upload and delete operations are done. There was an attempt to manage
sync state to ensure that file contents are identical but it's not doing
what I expected it to do yet.