Go back to one page for logged in and logged out. (#133)
We got into kindof a goofy spot having one route for logged in and one route for logged out. Revert back to the pattern of having GrindOlympiadsIndex handle both cases, which makes a smoother transition between logging in and out (i.e. don't randomly wind up on the wrong URL).